requirements.txt
# 使用するライブラリをここに列挙してください。

# サンプルプログラム
faiss-cpu>=1.13.2
google-genai>=1.56.0
python-dotenv


# 以下はサンプルです。
pypdf
unstructured
openai
supabase
fastapi
uvicorn